Win fixes

+++ bundledApps/WAIL.py
@@ -53,8 +53,18 @@ ssl._create_default_https_context = ssl._create_unverified_context
 WAIL_VERSION = "-1"
 INDEX_TIMER_SECONDS = 10.0
 
+wailPath = os.path.dirname(os.path.realpath(__file__))
+wailPath = wailPath.replace('\\bundledApps', '') # Fix for dev mode
+
+infoPlistPath = ""
+if 'darwin' in sys.platform:
+    infoPlistPath = "/Applications/WAIL.app/Contents/Info.plist"
+else:
+    infoPlistPath = wailPath + "\\build\\Info.plist"
+
+
 try:
-  with open ("/Applications/WAIL.app/Contents/Info.plist", "r") as myfile:
+  with open (infoPlistPath, "r") as myfile:
     data=myfile.read()
     m = re.search(r"<key>CFBundleShortVersionString</key>\n\t<string>(.*)</string>",
       data)
@@ -160,7 +170,8 @@ warcsFolder = ""
 tomcatPath = ""
 tomcatPathStart = ""
 tomcatPathStop = ""
-wailPath = os.path.dirname(os.path.realpath(__file__))
+memGatorPath = ""
+archivesJSON = ""
 fontSize = 8
 wailWindowSize = (400, 250)
 
@@ -198,13 +209,15 @@ elif sys.platform.startswith('win32'):
 
     aboutWindow_iconPath = wailPath + aboutWindow_iconPath
 
-    heritrixPath = "C:/WAIL/bundledApps/heritrix-3.2.0/"
-    heritrixBinPath = heritrixPath+"bin/heritrix.cmd"
-    heritrixJobPath = "C:\\WAIL\\jobs\\"
-    tomcatPath = "C:/WAIL/bundledApps/tomcat"
-    warcsFolder = wailPathPath + "archives"
-    tomcatPathStart = "C:/WAIL/support/catalina_start.bat"
-    tomcatPathStop = "C:/WAIL/support/catalina_stop.bat"
+    heritrixPath = wailPath + "\\bundledApps\\heritrix-3.2.0\\"
+    heritrixBinPath = heritrixPath + "\\bin\\heritrix.cmd"
+    heritrixJobPath = heritrixPath + "\\jobs\\"
+    tomcatPath = wailPath + "\\bundledApps\\tomcat"
+    warcsFolder = wailPath + "\\archives"
+    memGatorPath = wailPath + "\\bundledApps\\memgator-windows-amd64.exe"
+    archivesJSON = wailPath + "\\config\\archives.json"
+    tomcatPathStart = wailPath + "\\support\\catalina_start.bat"
+    tomcatPathStop = wailPath + "\\support\\catalina_stop.bat"
 ###############################
 # Tab Controller (Notebook)
 ###############################
@@ -277,6 +290,7 @@ class TabController(wx.Frame):
 class WAILGUIFrame_Basic(wx.Panel):
     def __init__(self, parent):
         wx.Panel.__init__(self, parent)
+        self.SetDoubleBuffered(True) # Forces Windows into composite mode for drawing
         self.uriLabel = wx.StaticText(self, -1, buttonLabel_uri, pos=(0, 5))
         self.uri = wx.TextCtrl(self, -1, pos=(34, 0), value=textLabel_defaultURI, size=(343, 25))
         self.archiveNowButton = wx.Button(self, -1, buttonLabel_archiveNow, pos=(280, 30))
@@ -306,13 +320,13 @@ class WAILGUIFrame_Basic(wx.Panel):
         self.uri.Bind(wx.EVT_KEY_UP, self.uriChanged) # Call memgator on URI change
     
     def setMementoCount(self, count):
-        if hasattr(self,'mementoStatus'):
+        if hasattr(self, 'mementoStatus'):
           self.mementoStatus.Destroy()
           self.mementoStatusPublicArchives.Destroy()
-          
+
         if count:
           memCountMsg = str(count) + " mementos available"
-          self.mementoStatus = wx.StaticText(self, -1, label=memCountMsg, pos=(105, 85), size=(150,20))
+          self.mementoStatus = wx.StaticText(self, -1, label=memCountMsg, pos=(105, 85), size=(150,20)) # Not updating UI on Windows
         else:
           self.mementoStatus = wx.StaticText(self, -1, label=msg_fetchingMementos, pos=(105, 85), size=(150,20))
           #italicFont = self.mementoStatus.GetFont().SetStyle(wx.ITALIC)
@@ -331,12 +345,13 @@ class WAILGUIFrame_Basic(wx.Panel):
         # TODO: Use CDXJ for counting the mementos
         out = check_output([memGatorPath, "-a", archivesJSON, self.uri.GetValue()])
         print "memgator command:"
-        print memGatorPath+" -a " + archivesJSON + self.uri.GetValue()
+        print memGatorPath + " -a " + archivesJSON + ' ' + self.uri.GetValue()
         
         # TODO: bug, on Gogo internet MemGator cannot hit aggregator, which results in 0 mementos, which MemGator throws exception
         
         # TODO: Once we are using the local web service, we can curl -I to get a 
-        self.setMementoCount(out.count("memento"))
+        self.setMementoCount(out.count("memento")) # UI not updated on Windows
+        print 'Setting memgator count to {0}'.format(out.count("memento"))
         # TODO: cache the TM
     def uriChanged(self, event):
        self.setMementoCount(None)
@@ -358,16 +373,30 @@ class WAILGUIFrame_Basic(wx.Panel):
         print "callback executed!"
 
     def ensureEnvironmentVariablesAreSet(self):
+        if 'darwin' not in sys.platform:
+            return # Allow windows to proceed w/o java checks for now.
+
         JAVA_HOME_defined = 'JAVA_HOME' in os.environ
         JRE_HOME_defined = 'JRE_HOME' in os.environ
         if not JAVA_HOME_defined or not JRE_HOME_defined:
+          jreHome = ''
+          javaHome = ''
+          jdkPath = ''
+          if 'darwin' in sys.platform:
+            jdkPath = "/Library/Java/JavaVirtualMachines/jdk1.7.0_79.jdk"
+            jreHome = "/Library/Java/JavaVirtualMachines/jdk1.7.0_79.jdk/Contents/Home"
+            javaHome = "/Library/Java/JavaVirtualMachines/jdk1.7.0_79.jdk/Contents/Home"
+          else: #Win, incomplete
+            os.environ['PATH'] # java8 does not use JRE_HOME, JAVA_HOME
+            
+          
           # Find java 1.7
           #/usr/libexec/java_home -v 1.7
-          jdkInstalled = os.path.isdir("/Library/Java/JavaVirtualMachines/jdk1.7.0_79.jdk")
+          jdkInstalled = os.path.isdir(jdkPath)
 
           if jdkInstalled:
-            os.environ["JAVA_HOME"] = "/Library/Java/JavaVirtualMachines/jdk1.7.0_79.jdk/Contents/Home"
-            os.environ["JRE_HOME"] = "/Library/Java/JavaVirtualMachines/jdk1.7.0_79.jdk/Contents/Home"
+            os.environ["JAVA_HOME"] = javaHome
+            os.environ["JRE_HOME"] = jreHome
             self.ensureEnvironmentVariablesAreSet()
           else:
             d = wx.MessageDialog(self, 'Java needs to be installed for Heritrix and Wayback', "Install now?", wx.YES_NO|wx.YES_DEFAULT|wx.ICON_QUESTION)
@@ -584,21 +613,30 @@ class WAILGUIFrame_Advanced(wx.Panel):
             self.status_wayback = wx.StaticText(self, 100, status,       (col1,    rowHeight*1),      cellSize)
         
         def getHeritrixVersion(self, abbr=True):
-            for file in os.listdir(heritrixPath + "lib/"):
+            htrixLibPath = heritrixPath + "lib/"
+            if 'darwin' not in sys.platform:
+                htrixLibPath = htrixLibPath.replace('/','\\')
+            for file in os.listdir(htrixLibPath):
               if file.startswith("heritrix-commons"):
                 regex = re.compile("commons-(.*)\.")
                 return regex.findall(file)[0]
 
         def getWaybackVersion(self):
-            for file in os.listdir(tomcatPath + "/webapps/lib/"):
+            tomcatLibPath = tomcatPath + "/webapps/lib/"
+            if 'darwin' not in sys.platform:
+                tomcatLibPath = tomcatLibPath.replace('/','\\')
+            for file in os.listdir(tomcatLibPath):
               if file.startswith("openwayback-core"):
                 regex = re.compile("core-(.*)\.")
                 return regex.findall(file)[0]
 
         def getTomcatVersion(self):
         #Apache Tomcat Version 7.0.30
-            if not os.path.exists(tomcatPath+"/RELEASE-NOTES"): return "?"
-            f = open(tomcatPath+"/RELEASE-NOTES",'r')
+            releaseNotesPath = tomcatPath + '/RELEASE-NOTES'
+            if 'darwin' not in sys.platform:
+                releaseNotesPath = releaseNotesPath.replace('/','\\')
+            if not os.path.exists(releaseNotesPath): return "?"
+            f = open(releaseNotesPath,'r')
             version = ""
             for line in f.readlines():
                 if "Apache Tomcat Version " in line:
@@ -705,10 +743,11 @@ class WAILGUIFrame_Advanced(wx.Panel):
                     #time.sleep(3)
                     #self.openWaybackInBrowser(None)
         def openWaybackConfiguration(self,button):
-            filepath = tomcatPath+"/webapps/ROOT/WEB-INF/wayback.xml"
+            filepath = tomcatPath + "/webapps/ROOT/WEB-INF/wayback.xml"
             if sys.platform.startswith('darwin'):
              subprocess.call(('open', filepath))
             elif os.name == 'nt':
+             filepath = filepath.replace('/','\\')
              os.startfile(filepath)
             elif os.name == 'posix':
              subprocess.call(('xdg-open', filepath))
@@ -830,6 +869,7 @@ class WAILGUIFrame_Advanced(wx.Panel):
             if sys.platform.startswith('darwin'):
                 subprocess.call(('open', file))
             elif os.name == 'nt':
+                file = file.replace('/','\\')
                 os.startfile(file)
             elif os.name == 'posix':
                 subprocess.call(('xdg-open', file))
@@ -1095,7 +1135,7 @@ class Service():
                 return True
            # if hasattr(e, 
 
-            print "Failed to access " + self.__class__.__name__+" service at " + self.uri
+            print "Failed to access " + self.__class__.__name__  +" service at " + self.uri
             return False
         except:
             print "Some other error occurred in trying to check service accessibility."
@@ -1109,7 +1149,7 @@ class Wayback(Service):
 
     def fixAsync(self, cb=None):
         mainAppWindow.advConfig.generalPanel.updateServiceStatuses("wayback","FIXING")
-        cmd = tomcatPathStart;
+        cmd = tomcatPathStart
         ret = subprocess.Popen(cmd)
         time.sleep(3)
         wx.CallAfter(mainAppWindow.advConfig.generalPanel.updateServiceStatuses)
@@ -1133,8 +1173,15 @@ class Wayback(Service):
         self.generateCDX()
 
     def generatePathIndex(self):
-        dest = "/Applications/WAIL.app/config/path-index.txt"
-        warcsPath = "/Applications/WAIL.app/archives/"
+        wailPath = '/Applications/WAIL.app'
+        if 'darwin' not in sys.platform:
+            wailPath = 'C:\wail'
+
+        dest = wailPath + "/config/path-index.txt"
+        warcsPath = wailPath + "/archives/"
+        if 'darwin' not in sys.platform:
+            dest = dest.replace('/','\\')
+            warcsPath = warcsPath.replace('/','\\')
         
         outputContents = ""
         for file in listdir(warcsPath):
@@ -1149,21 +1196,40 @@ class Wayback(Service):
     
     def generateCDX(self):
         #/Applications/WAIL.app/bundledApps/tomcat/webapps/bin/cdx-indexer (file) (file.cdx)
-        dest = "/Applications/WAIL.app/config/path-index.txt"
-        warcsPath = "/Applications/WAIL.app/archives/"
+        wailRoot = '/Applications/WAIL.app'
+        if 'darwin' not in sys.platform:
+            wailRoot = 'C:\wail'
+        dest = wailRoot + "/config/path-index.txt"
+        warcsPath = wailRoot + "/archives/"
+        cdxFilePathPre = wailRoot + "/archiveIndexes/"
+        cdxIndexerPath = wailRoot +  "/bundledApps/tomcat/webapps/bin/cdx-indexer"
+
+        if 'darwin' not in sys.platform:
+            dest = dest.replace('/','\\')
+            warcsPath = warcsPath.replace('/','\\')
+            cdxFilePathPre = cdxFilePathPre.replace('/','\\')
+            cdxIndexerPath = cdxIndexerPath.replace('/','\\')
         
         outputContents = ""
         for file in listdir(warcsPath):
             if file.endswith(".warc"):
-              cdxFilePath = "/Applications/WAIL.app/archiveIndexes/" + file.replace('.warc','.cdx')
-              process = subprocess.Popen(["/Applications/WAIL.app/bundledApps/tomcat/webapps/bin/cdx-indexer",join(warcsPath,file),cdxFilePath], stdout=PIPE, stderr=PIPE)
+              cdxFilePath = cdxFilePathPre + file.replace('.warc','.cdx')
+              process = subprocess.Popen([cdxIndexerPath,join(warcsPath,file),cdxFilePath], stdout=PIPE, stderr=PIPE)
               stdout, stderr = process.communicate()
         
         # Combine CDX files
-        filenames = glob.glob("/Applications/WAIL.app/archiveIndexes/*.cdx")
+        allCDXesPath = wailRoot + "/archiveIndexes/*.cdx"
+        if 'darwin' not in sys.platform:
+            allCDXesPath = allCDXesPath.replace('/','\\')
+
+        filenames = glob.glob(allCDXesPath)
         cdxHeaderIncluded = False
         print "CDX files generated for each WARC, merging..."
-        with open('/Applications/WAIL.app/archiveIndexes/combined_unsorted.cdxt', 'w') as outfile:
+        unsortedPath = wailRoot + '/archiveIndexes/combined_unsorted.cdxt' # Is cdxt the right filename?
+        if 'darwin' not in sys.platform:
+            unsortedPath = unsortedPath.replace('/','\\')
+
+        with open(unsortedPath, 'w') as outfile:
             for fname in filenames:
                 with open(fname) as infile:
                     for i,line in enumerate(infile):
@@ -1173,16 +1239,18 @@ class Wayback(Service):
                         outfile.write(line)
                         cdxHeaderIncluded = True
         print "Done merging CDX files, removing old source CDX files."
-        filelist = glob.glob("/Applications/WAIL.app/archiveIndexes/*.cdx")
+        filelist = glob.glob(allCDXesPath)
         for f in filelist:
             os.remove(f)
+
+        # TODO: fix cdx sorting in Windows
+        if 'darwin' in sys.platform:
+          print "Sorting CDX entries"
+          os.system("export LC_ALL=C; sort -u /Applications/WAIL.app/archiveIndexes/combined_unsorted.cdxt > /Applications/WAIL.app/archiveIndexes/index.cdx")
+          print "Removing unsorted temp file"
+          os.remove("/Applications/WAIL.app/archiveIndexes/combined_unsorted.cdxt")
         
-        print "Sorting CDX entries"
-        os.system("export LC_ALL=C; sort -u /Applications/WAIL.app/archiveIndexes/combined_unsorted.cdxt > /Applications/WAIL.app/archiveIndexes/index.cdx")
-        print "Removing unsorted temp file"
-        os.remove("/Applications/WAIL.app/archiveIndexes/combined_unsorted.cdxt")
-        
-        print "Done creating sorted CDX file!"
+          print "Done creating sorted CDX file!"
         
         # Queue next iteration of indexing
         if mainAppWindow.indexingTimer:
@@ -1219,7 +1287,11 @@ class Heritrix(Service):
         for launch in launches:
             #print heritrixJobPath+jobId+"/"+launch+"/logs/progress-statistics.log"
             print heritrixJobPath+jobId+"/"+launch+"/logs/progress-statistics.log"
-            lastLine = tail(heritrixJobPath+jobId+"/"+launch+"/logs/progress-statistics.log")
+            progressLogFilePath = heritrixJobPath + jobId + "/" + launch + "/logs/progress-statistics.log"
+            if 'darwin' not in sys.platform:
+                progressLogFilePath = progressLogFilePath.replace('/','\\')
+
+            lastLine = tail(progressLogFilePath)
 
             ll = lastLine[0].replace(" ","|")
             logData = re.sub(r'[|]+', '|', ll).split("|")
@@ -1262,7 +1334,7 @@ class HeritrixJob:
             beansFilePath += "\\"
         else:
             beansFilePath += "/"
-        with open(beansFilePath+"crawler-beans.cxml","w") as f:
+        with open(beansFilePath + "crawler-beans.cxml","w") as f:
             f.write(self.sampleXML)
             #print beansFilePath+"crawler-beans.cxml"
 
@@ -2052,7 +2124,7 @@ class UpdateSoftwareWindow(wx.Frame):
         self.currentVersion_heritrix = self.getHeritrixVersion()
         self.currentVersion_wayback = self.getWaybackVersion()
         
-        packages = self.updateJSONData['packages'];
+        packages = self.updateJSONData['packages']
         for package in packages:
           if package['name'] == 'heritrix-wail':
             self.latestVersion_heritrix = package['version']
